+++ laravel/blade.php
@@ -9,6 +9,7 @@ class Blade {
 	 */
 	protected static $compilers = array(
 		'layouts',
+		'comments',
 		'echos',
 		'forelse',
 		'empty',
@@ -16,6 +17,8 @@ class Blade {
 		'structure_openings',
 		'structure_closings',
 		'else',
+		'unless',
+		'endunless',
 		'includes',
 		'render_each',
 		'render',
@@ -151,6 +154,19 @@ class Blade {
 	}
 
 	/**
+	 * Rewrites Blade comments into PHP comments.
+	 *
+	 * @param  string  $value
+	 * @return string
+	 */
+	protected static function compile_comments($value)
+	{
+		$value = preg_replace('/\{\{--(.+?)(--\}\})?\n/', "<?php // $1 ?>", $value);
+
+		return preg_replace('/\{\{--((.|\s)*?)--\}\}/', "<?php /* $1 */ ?>\n", $value);
+	}
+
+	/**
 	 * Rewrites Blade echo statements into PHP echo statements.
 	 *
 	 * @param  string  $value
@@ -255,6 +271,30 @@ class Blade {
 	}
 
 	/**
+	 * Rewrites Blade "unless" statements into valid PHP.
+	 *
+	 * @param  string  $value
+	 * @return string
+	 */
+	protected static function compile_unless($value)
+	{
+		$pattern = '/(\s*)@unless(\s*\(.*\))/';
+
+		return preg_replace($pattern, '$1<?php if( ! ($2)): ?>', $value);
+	}
+
+	/**
+	 * Rewrites Blade "unless" endings into valid PHP.
+	 *
+	 * @param  string  $value
+	 * @return string
+	 */
+	protected static function compile_endunless($value)
+	{
+		return str_replace('@endunless', '<?php endif; ?>', $value);
+	}
+
+	/**
 	 * Rewrites Blade @include statements into valid PHP.
 	 *
 	 * @param  string  $value
@@ -264,7 +304,7 @@ class Blade {
 	{
 		$pattern = static::matcher('include');
 
-		return preg_replace($pattern, '$1<?php echo view$2->with(get_defined_vars()); ?>', $value);
+		return preg_replace($pattern, '$1<?php echo view$2->with(get_defined_vars())->render(); ?>', $value);
 	}
 
 	/**