+++ laravel/database/eloquent/model.php
@@ -56,6 +56,13 @@ abstract class Model {
 	public static $accessible;
 
 	/**
+	 * The attributes that should be excluded from to_array.
+	 *
+	 * @var array
+	 */
+	public static $hidden = array();
+
+	/**
 	 * Indicates if the model has update and creation timestamps.
 	 *
 	 * @var bool
@@ -186,7 +193,7 @@ abstract class Model {
 	{
 		$model = new static(array(), true);
 
-		if (static::$timestamps) $attributes['updated_at'] = $model->get_timestamp();
+		if (static::$timestamps) $attributes['updated_at'] = new \DateTime;
 
 		return $model->query()->where($model->key(), '=', $id)->update($attributes);
 	}
@@ -398,19 +405,9 @@ abstract class Model {
 	 */
 	protected function timestamp()
 	{
-		$this->updated_at = $this->get_timestamp();
+		$this->updated_at = new \DateTime;
 
 		if ( ! $this->exists) $this->created_at = $this->updated_at;
-	}
-
-	/**
-	 * Get the current timestamp in its storable form.
-	 *
-	 * @return mixed
-	 */
-	public function get_timestamp()
-	{
-		return date('Y-m-d H:i:s');
 	}
 
 	/**
@@ -475,7 +472,17 @@ abstract class Model {
 	 */
 	public function get_dirty()
 	{
-		return array_diff_assoc($this->attributes, $this->original);
+		$dirty = array();
+
+		foreach ($this->attributes as $key => $value)
+		{
+			if ( ! isset($this->original[$key]) or $value !== $this->original[$key])
+			{
+				$dirty[$key] = $value;
+			}
+		}
+
+		return $dirty;
 	}
 
 	/**
@@ -531,6 +538,51 @@ abstract class Model {
 		unset($this->original[$key]);
 
 		unset($this->attributes[$key]);
+	}
+
+	/**
+	 * Get the model attributes and relationships in array form.
+	 *
+	 * @return array
+	 */
+	public function to_array()
+	{
+		$attributes = array();
+
+		// First we need to gather all of the regular attributes. If the attribute
+		// exists in the array of "hidden" attributes, it will not be added to
+		// the array so we can easily exclude things like passwords, etc.
+		foreach (array_keys($this->attributes) as $attribute)
+		{
+			if ( ! in_array($attribute, static::$hidden))
+			{
+				$attributes[$attribute] = $this->$attribute;
+			}
+		}
+
+		foreach ($this->relationships as $name => $models)
+		{
+			// If the relationship is not a "to-many" relationship, we can just
+			// to_array the related model and add it as an attribute to the
+			// array of existing regular attributes we gathered.
+			if ( ! is_array($models))
+			{
+				$attributes[$name] = $models->to_array();
+			}
+
+			// If the relationship is a "to-many" relationship we need to spin
+			// through each of the related models and add each one with the
+			// to_array method, keying them both by name and ID.
+			else
+			{
+				foreach ($models as $id => $model)
+				{
+					$attributes[$name][$id] = $model->to_array();
+				}
+			}
+		}
+
+		return $attributes;
 	}
 
 	/**
@@ -598,6 +650,8 @@ abstract class Model {
 		{
 			if (array_key_exists($key, $this->$source)) return true;
 		}
+		
+		if (method_exists($this, $key)) return true;
 	}
 
 	/**
@@ -623,10 +677,12 @@ abstract class Model {
 	 */
 	public function __call($method, $parameters)
 	{
+		$meta = array('key', 'table', 'connection', 'sequence', 'per_page', 'timestamps');
+
 		// If the method is actually the name of a static property on the model we'll
 		// return the value of the static property. This makes it convenient for
 		// relationships to access these values off of the instances.
-		if (in_array($method, array('key', 'table', 'connection', 'sequence', 'per_page')))
+		if (in_array($method, $meta))
 		{
 			return static::$$method;
 		}
@@ -644,11 +700,11 @@ abstract class Model {
 		// to perform the appropriate action based on the method.
 		if (starts_with($method, 'get_'))
 		{
-			return $this->attributes[substr($method, 4)];
+			return $this->get_attribute(substr($method, 4));
 		}
 		elseif (starts_with($method, 'set_'))
 		{
-			$this->attributes[substr($method, 4)] = $parameters[0];
+			$this->set_attribute(substr($method, 4), $parameters[0]);
 		}
 
 		// Finally we will assume that the method is actually the beginning of a