+++ laravel/database/schema/grammars/postgres.php
@@ -368,7 +368,7 @@ class Postgres extends Grammar {
 	 */
 	protected function type_date(Fluent $column)
 	{
-		return 'TIMESTAMP';
+		return 'TIMESTAMP(0) WITHOUT TIME ZONE';
 	}
 
 	/**