+++ laravel/documentation/changes.md
@@ -1,7 +1,9 @@
-## Laravel Change Log
+# Laravel Change Log
 
 ## Contents
 
+- [Laravel 3.2](#3.2)
+- [Upgrading From 3.1](#upgrade-3.2)
 - [Laravel 3.1.7](#3.1.7)
 - [Upgrading From 3.1.6](#upgrade-3.1.7)
 - [Laravel 3.1.6](#3.1.6)
@@ -18,6 +20,34 @@
 - [Upgrading From 3.1](#upgrade-3.1.1)
 - [Laravel 3.1](#3.1)
 - [Upgrading From 3.0](#upgrade-3.1)
+
+<a name="3.2"></a>
+## Laravel 3.2
+
+- [Added `to_array` method to the base Eloquent model](/docs/database/eloquent#to-array).
+- [Added `$hidden` static variable to the base Eloquent model](/docs/database/eloquent#to-array).
+- [Added `sync` method to has\_many\_and\_belongs\_to Eloquent relationship](/docs/database/eloquent#sync-method).
+- [Added `save` method to has\_many Eloquent relationship](/docs/database/eloquent#has-many-save).
+- [Added `unless` structure to Blade template engine](/docs/views/templating#blade-unless).
+- [Added Blade comments](/docs/views/templating#blade-comments).
+- [Added simpler environment management](/docs/install#environments).
+- Added support for DateTime database query bindings.
+- Migrated to the Symfony HttpFoundation component for core request / response handling.
+- Fixed the passing of strings into the `Input::except` method.
+- Fixed replacement of optional parameters in `URL::transpose` method.
+- Improved `update` handling on `Has_Many` and `Has_One` relationships.
+- Improved View performance by only loading contents from file once.
+- Fix handling of URLs beginning with has in `URL::to`.
+- Fix the resolution of unset Eloquent attributes.
+- Allows pivot table timestamps to be disabled.
+- Made the `get_timestamp` Eloquent method static.
+- `Request::secure` now takes `application.ssl` configuration option into consideration.
+- Simplified the `paths.php` file.
+
+<a name="upgrade-3.2"></a>
+## Upgrading From 3.1
+
+- Replace the **laravel** folder.
 
 <a name="3.1.7"></a>
 ## Laravel 3.1.7